feat(overview): add overview block (#317)
