+++ docs/app/app.module.ts
@@ -23,7 +23,6 @@ import { DocsService } from './docs/docs.service';
 import { NgdDocsComponent } from './docs/docs.component';
 import { NgdPageComponent } from './docs/page/page.component';
 import { NgdMarkdownComponent } from './docs/page/blocks/ngd-markdown-block.component';
-import { NgdDescriptionBlockComponent } from './docs/page/blocks/basic-blocks/ngd-description-block.component';
 import { NgdExamplesBlockComponent } from './docs/page/blocks/basic-blocks/ngd-examples-block.component';
 import { NgdPropsBlockComponent } from './docs/page/blocks/basic-blocks/ngd-props-block.component';
 import { NgdMethodsBlockComponent } from './docs/page/blocks/basic-blocks/ngd-methods-block.component';
@@ -35,6 +34,7 @@ import { NgdHighlighterComponent } from './docs/utils/code-highlighter.component
 import { NgdHeaderComponent } from './components/header/ngd-header.component';
 import { NgdFooterComponent } from './components/footer/ngd-footer.component';
 import { NgdStylesBlockComponent } from './docs/page/blocks/basic-blocks/ngd-styles-block.component';
+import { NgdOverviewBlockComponent } from './docs/page/blocks/basic-blocks/ngd-overview-block.component';
 import { NgdComponentBlockComponent } from './docs/page/blocks/ngd-component-block.component';
 import { NgdFragmentDirective } from './docs/utils/ngd-fragment.directive';
 import { NgdThemeComponent } from './docs/page/blocks/ngd-theme-block.component';
@@ -65,7 +65,6 @@ import { BlockHelperService } from './docs/utils/block-helper.service';
     NgdDocsComponent,
     NgdPageComponent,
     NgdMarkdownComponent,
-    NgdDescriptionBlockComponent,
     NgdExamplesBlockComponent,
     NgdPropsBlockComponent,
     NgdMethodsBlockComponent,
@@ -79,6 +78,7 @@ import { BlockHelperService } from './docs/utils/block-helper.service';
     NgdThemeComponent,
     NgdBlockComponent,
     NgdTabbedBlockComponent,
+    NgdOverviewBlockComponent,
     NgdSassPropValueDirective,
   ],
   providers: [
