+++ docs/app/docs/utils/block-helper.service.ts
@@ -2,21 +2,6 @@ import { Injectable } from '@angular/core';
 
 @Injectable()
 export class BlockHelperService {
-
-  hasDescription(component): boolean {
-    return component && (
-      component.name ||
-      component.shortDescription ||
-      component.description
-    );
-  }
-
-  hasExamples(component): boolean {
-    return component &&
-      component.examples &&
-      component.examples.length > 0;
-  }
-
   hasTheme(component): boolean {
     return component.styles &&
       component.styles.length > 0;
@@ -33,5 +18,9 @@ export class BlockHelperService {
       component.methods &&
       component.methods.length > 0 &&
       component.methods.some(method => method.shortDescription || method.description);
+  }
+
+  hasOverview(component): boolean {
+    return component && component.overview && component.overview.length;
   }
 }
